fix: [EL-4910] Fixed when clicking pin/unpin and dot menu triggered bucket expansion
